Get rid of packages/domain-picker and domain suggestions data store in favor of React Query fetching
