+++ client/package.json
@@ -36,7 +36,6 @@
 		"@automattic/data-stores": "workspace:^",
 		"@automattic/design-picker": "workspace:^",
 		"@automattic/design-preview": "workspace:^",
-		"@automattic/domain-picker": "workspace:^",
 		"@automattic/domain-search": "workspace:^",
 		"@automattic/domains-table": "workspace:^",
 		"@automattic/explat-client": "workspace:^",
