--- packages/data-stores/src/domain-suggestions/actions.ts
@@ -1,47 +1,0 @@
-import type { DomainSuggestion, DomainCategory, DomainAvailability } from './types';
-import type { DomainSuggestionQuery } from '@automattic/domain-search';
-
-export const receiveCategories = ( categories: DomainCategory[] ) =>
-	( {
-		type: 'RECEIVE_CATEGORIES',
-		categories,
-	} ) as const;
-
-export const fetchDomainSuggestions = () =>
-	( {
-		type: 'FETCH_DOMAIN_SUGGESTIONS',
-		timeStamp: Date.now(),
-	} ) as const;
-
-export const receiveDomainAvailability = ( domainName: string, availability: DomainAvailability ) =>
-	( {
-		type: 'RECEIVE_DOMAIN_AVAILABILITY',
-		domainName,
-		availability,
-	} ) as const;
-
-export const receiveDomainSuggestionsSuccess = (
-	queryObject: DomainSuggestionQuery,
-	suggestions: DomainSuggestion[] | undefined
-) =>
-	( {
-		type: 'RECEIVE_DOMAIN_SUGGESTIONS_SUCCESS',
-		queryObject,
-		suggestions,
-		timeStamp: Date.now(),
-	} ) as const;
-
-export const receiveDomainSuggestionsError = ( errorMessage: string ) =>
-	( {
-		type: 'RECEIVE_DOMAIN_SUGGESTIONS_ERROR',
-		errorMessage,
-		timeStamp: Date.now(),
-	} ) as const;
-
-export type Action = ReturnType<
-	| typeof receiveCategories
-	| typeof fetchDomainSuggestions
-	| typeof receiveDomainSuggestionsSuccess
-	| typeof receiveDomainSuggestionsError
-	| typeof receiveDomainAvailability
->;
