--- packages/data-stores/src/domain-suggestions/constants.ts
@@ -1,8 +1,0 @@
-export const STORE_KEY = 'automattic/domains/suggestions';
-
-export enum DataStatus {
-	Failure = 'failure',
-	Pending = 'pending',
-	Success = 'success',
-	Uninitialized = 'uninitialized',
-}
