--- packages/domain-picker/.eslintrc.js
@@ -1,10 +1,0 @@
-module.exports = {
-	overrides: [
-		{
-			files: [ '**/__tests__/**/*', '**/__mocks__/**/*' ],
-			rules: {
-				'jest/no-mocks-import': 'off',
-			},
-		},
-	],
-};
