--- packages/domain-picker/jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-	preset: '../../test/packages/jest-preset.js',
-	testEnvironment: 'jsdom',
-	testMatch: [ '<rootDir>/**/__tests__/**/*.[jt]s?(x)', '!**/.eslintrc.*' ],
-	transformIgnorePatterns: [ 'node_modules/(?!gridicons)(?!.*\\.svg)' ],
-};
