--- packages/domain-picker/package.json
@@ -1,73 +1,0 @@
-{
-	"name": "@automattic/domain-picker",
-	"version": "1.0.0-alpha.0",
-	"description": "Domain picker.",
-	"homepage": "https://github.com/Automattic/wp-calypso",
-	"license": "GPL-2.0-or-later",
-	"author": "Automattic Inc.",
-	"main": "dist/cjs/index.js",
-	"module": "dist/esm/index.js",
-	"calypso:src": "src/index.tsx",
-	"exports": {
-		".": {
-			"calypso:src": "./src/index.tsx",
-			"types": "./dist/types/index.d.ts",
-			"import": "./dist/esm/index.js",
-			"require": "./dist/cjs/index.js"
-		}
-	},
-	"sideEffects": [
-		"*.css",
-		"*.scss"
-	],
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/Automattic/wp-calypso.git",
-		"directory": "packages/domain-picker"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"bugs": "https://github.com/Automattic/wp-calypso/issues",
-	"types": "dist/types",
-	"scripts": {
-		"clean": "tsc --build ./tsconfig.json ./tsconfig-cjs.json --clean && rm -rf dist",
-		"build": "tsc --build ./tsconfig.json ./tsconfig-cjs.json && copy-assets",
-		"prepack": "yarn run clean && yarn run build",
-		"watch": "tsc --build ./tsconfig.json --watch"
-	},
-	"dependencies": {
-		"@automattic/calypso-analytics": "workspace:^",
-		"@automattic/calypso-color-schemes": "workspace:^",
-		"@automattic/data-stores": "workspace:^",
-		"@automattic/i18n-utils": "workspace:^",
-		"@automattic/onboarding": "workspace:^",
-		"@automattic/typography": "workspace:^",
-		"@wordpress/base-styles": "^5.23.0",
-		"@wordpress/components": "^29.9.0",
-		"@wordpress/compose": "^7.23.0",
-		"@wordpress/icons": "^10.23.0",
-		"@wordpress/react-i18n": "^4.23.0",
-		"clsx": "^2.1.1",
-		"lodash": "^4.17.21",
-		"tslib": "^2.8.1",
-		"use-debounce": "^3.1.0"
-	},
-	"devDependencies": {
-		"@automattic/calypso-typescript-config": "workspace:^",
-		"@testing-library/dom": "^10.4.0",
-		"@testing-library/jest-dom": "^6.6.3",
-		"@testing-library/react": "^16.3.0",
-		"react": "^18.3.1",
-		"react-dom": "^18.3.1",
-		"typescript": "^5.8.3"
-	},
-	"peerDependencies": {
-		"@wordpress/data": "^10.23.0",
-		"@wordpress/element": "^6.23.0",
-		"@wordpress/i18n": "^5.23.0",
-		"react": "^18.3.1",
-		"redux": "^5.0.1"
-	},
-	"private": true
-}
