--- packages/domain-picker/src/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './use-domain-suggestions';
