--- packages/domain-picker/src/index.tsx
@@ -1,1 +1,0 @@
-export { useDomainSuggestions } from './hooks';
