--- packages/domain-picker/src/types.d.ts
@@ -1,1 +1,0 @@
-declare const __i18n_text_domain__: string;
