--- packages/domain-picker/tsconfig.json
@@ -1,16 +1,0 @@
-{
-	"extends": "@automattic/calypso-typescript-config/ts-package.json",
-	"compilerOptions": {
-		"declarationDir": "dist/types",
-		"outDir": "dist/esm",
-		"rootDir": "src"
-	},
-	"include": [ "src" ],
-	"exclude": [ "**/__tests__/*", "**/__mocks__/*" ],
-	"references": [
-		{ "path": "../data-stores" },
-		{ "path": "../calypso-analytics" },
-		{ "path": "../i18n-utils" },
-		{ "path": "../onboarding" }
-	]
-}
